Better network management for product fetching Removed productMonitorWorker and ProductWatcherService Up version to 2.6.1
