+++ bot/manifest.json
@@ -2,7 +2,7 @@
   "manifest_version": 2,
   "name": "Supreme Auto-Checkout bot",
   "description": "Supreme Open-Source add to cart bot",
-  "version": "2.6",
+  "version": "2.6.1",
   "background": {
     "scripts": ["background.js"]
   },
