+++ bot/src/app/version.js
@@ -1,1 +1,1 @@
-export default '2.6';
+export default '2.6.1';
