+++ bot/src/services/supreme/ProductsService.js
@@ -40,3 +40,15 @@ export default class ProductsService {
     });
   }
 }
+
+chrome.webRequest.onBeforeSendHeaders.addListener(
+  function (details) {
+    const userAgent = details.requestHeaders.filter(x => x.name === 'User-Agent')[0];
+    if (userAgent) {
+      userAgent.value = 'Mozilla/5.0 (iPhone; CPU iPhone OS 9_3_3 like Mac OS X) AppleWebKit/601.1.46 (KHTML, like Gecko) Mobile/13G34';
+    }
+    return { requestHeaders: details.requestHeaders };
+  },
+  { urls: ['*://*.supremenewyork.com/mobile/*', '*://*.supremenewyork.com/mobile_stock.json*', 'http://www.supremenewyork.com/shop/*.json'] },
+  ['blocking', 'requestHeaders'],
+);
